added logic for 'favorites' playlist

+++ app/src/main/java/com/woolenstorm/musicplayer/ui/AppViewModel.kt
@@ -1,11 +1,13 @@
 package com.woolenstorm.musicplayer.ui
 
+import android.content.ContentProvider
 import android.content.Context
 import android.content.Intent
 import android.media.AudioAttributes
 import android.util.Log
 import androidx.annotation.DrawableRes
 import androidx.compose.runtime.*
+import androidx.compose.ui.res.stringResource
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.*
 import androidx.lifecycle.ViewModelProvider.AndroidViewModelFactory.Companion.APPLICATION_KEY
@@ -16,6 +18,7 @@ import com.woolenstorm.musicplayer.R
 import com.woolenstorm.musicplayer.communication.PlaybackService
 import com.woolenstorm.musicplayer.data.SongsRepository
 import com.woolenstorm.musicplayer.model.*
+import com.woolenstorm.musicplayer.utils.FAVORITES_PLAYLIST
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
 import kotlin.random.Random
@@ -31,6 +34,7 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
     val navigationType = mutableStateOf(NavigationType.BottomNavigation)
     val currentPosition = mutableStateOf(if (mediaPlayer.currentPosition < mediaPlayer.duration) mediaPlayer.currentPosition.toFloat() else 0f)
     val uiState = songsRepository.uiState
+    var favorites = songsRepository.favorites
     val navigationItemList = listOf(
         NavigationItemContent(
             type = CurrentScreen.Songs,
@@ -44,7 +48,7 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
     var isSearching = mutableStateOf(false)
     var songs = songsRepository.songs.toMutableStateList()
         private set
-    var backlog = songsRepository.backlog
+    private var backlog = songsRepository.backlog
     val playlists = database.playlistDao().getAll().map {
         PlaylistsUiState(it)
     }.stateIn(
@@ -65,6 +69,15 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
         startProgressSlider()
     }
 
+    fun checkIfFavoritesExist() {
+        viewModelScope.launch {
+            delay(2000)
+            playlists.value.itemList
+                .find { !it.canBeDeleted }
+                ?: createPlaylist(name = FAVORITES_PLAYLIST, canBeDeleted = false)
+        }
+    }
+
 
     fun updateCurrentScreen(newScreen: CurrentScreen) {
         currentScreen.value = newScreen
@@ -79,12 +92,12 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
         navigationType.value = newNavigationType
     }
 
-    suspend fun createPlaylist(name: String) {
-        database.playlistDao().insertPlaylist(Playlist(name = name))
+    suspend fun createPlaylist(name: String, canBeDeleted: Boolean = true) {
+        database.playlistDao().insertPlaylist(Playlist(name = name, canBeDeleted = canBeDeleted))
     }
 
     suspend fun deletePlaylist(playlist: Playlist) {
-        database.playlistDao().delete(playlist)
+        if (playlist.canBeDeleted) database.playlistDao().delete(playlist)
     }
 
     suspend fun updatePlaylist(playlist: Playlist) {
@@ -115,7 +128,8 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
         isHomeScreen: Boolean? = null,
         currentPosition: Float? = null,
         isExpanded: Boolean? = null,
-        playlistId: Int? = null
+        playlistId: Int? = null,
+        isFavored: Boolean? = null
     ) {
         songsRepository.updateUiState(
             song = song,
@@ -128,7 +142,8 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
             isHomeScreen = isHomeScreen,
             currentPosition = currentPosition,
             isExpanded = isExpanded,
-            playlistId = playlistId
+            playlistId = playlistId,
+            isFavored = isFavored
         )
     }
 
@@ -137,7 +152,8 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
         updateUiState(
             currentIndex = songs.indexOf(song),
             isSongChosen = true,
-            isHomeScreen = navigationType.value == NavigationType.NavigationRail
+            isHomeScreen = navigationType.value == NavigationType.NavigationRail,
+            isFavored = checkIfFavorite(song)
         )
         when {
             song == uiState.value.song && uiState.value.isPlaying -> {}
@@ -174,6 +190,31 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
         createNotification(context)
     }
 
+    fun onToggleFavored(context: Context) {
+
+        updateUiState(isFavored = !uiState.value.isFavored)
+        playlists.value.itemList
+            .find { !it.canBeDeleted }
+            ?.let {
+                val songsIds = emptyList<Long>().toMutableList()
+                songsIds.addAll(it.songsIds)
+
+                if (!uiState.value.isFavored) {
+                    songsIds.remove(uiState.value.song.id)
+                } else {
+                    songsIds.add(uiState.value.song.id)
+                }
+                val newPlaylist = it.copy(
+                    songsIds = songsIds
+                )
+
+                viewModelScope.launch {
+                    updatePlaylist(newPlaylist)
+                }
+            }
+        createNotification(context = context)
+    }
+
     fun nextSong(context: Context) {
         if (backlog.empty()) {
             backlog.push(uiState.value.song)
@@ -188,13 +229,18 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
             updateUiState(
                 song = currSongs[newIndex],
                 currentIndex = newIndex,
-                currentPosition = 0f
+                currentPosition = 0f,
+                isFavored = checkIfFavorite(currSongs[newIndex])
             )
             backlog.push(currSongs[newIndex])
             cancel(context)
             play(context)
         }
 
+    }
+
+    private fun checkIfFavorite(song: Song): Boolean {
+        return playlists.value.itemList.find { !it.canBeDeleted }?.songsIds?.contains(song.id) ?: false
     }
 
     fun previousSong(context: Context) {
@@ -208,7 +254,8 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
                 updateUiState(
                     song = lastSong,
                     currentIndex = currSongs.indexOf(lastSong),
-                    currentPosition = 0f
+                    currentPosition = 0f,
+                    isFavored = checkIfFavorite(lastSong)
                 )
             } else {
                 val newIndex = if (uiState.value.currentIndex <= 0) currSongs.size - 1 else uiState.value.currentIndex - 1
@@ -216,7 +263,8 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
                 updateUiState(
                     song = currSongs[newIndex],
                     currentIndex = newIndex,
-                    currentPosition = 0f
+                    currentPosition = 0f,
+                    isFavored = checkIfFavorite(currSongs[newIndex])
                 )
             }
             cancel(context)
@@ -234,6 +282,7 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
     }
 
     private fun play(context: Context) {
+
         cancel(context)
         mediaPlayer.apply {
             setAudioAttributes(
@@ -249,6 +298,7 @@ class AppViewModel(private val songsRepository: SongsRepository) : ViewModel() {
             start()
         }
         updateUiState(isPlaying = true)
+        songsRepository.favorites = playlists.value.itemList.find { !it.canBeDeleted }
         startProgressSlider()
         createNotification(context)
     }
